nil checks on (typed) interfaces are an abomination
An interface is composed of both the type and value. When oops.Trace
returned the pointer *TraceError instead of the interface error the type
information was passed along. Unfortunately that meant when passing
*TraceError to a function that takes error values (e.g. Chain) the error
interface had non-empty type information. This meant that err == nil
would return false in such functions because interface{*oops.Trace, nil}
is not the same as interface{nil, nil}. This is unfortunate since it
means we can't follow the usual paradigm of returning concrete types and
accepting interfaces... Or at least we can't and still rely on the
widely accepted `err == nil` practice. The options for not checking nil
in this way are all deficient in some way (performance, convention).

Further reading for the curious:

https://dave.cheney.net/2017/08/09/typed-nils-in-go-2

Changing all the functions to return the interface error instead of the
concrete type was deemed to least problematic way to resolve the
situation. Now the naive check err == nil works as expected. Eventually
maybe Go2 will provide a solution to this situation that is less
painful.