Rework Namespaces
This reworks namespaces to make them easier to use. Now most of the
library functions are available as methods on the namespace. Namespaces
now have roughly the same UX as the package itself.

TraceSkipInternal is now also exported to make it consistent how many
frames should be skipped to avoid reporting frames internal to the oops
package itself.