trace: capture all callers
The original intention was to capture all callers, but I misunderstood
the API contract and it only captured the first N (where N was 10).
This exposes and sets a higher chunk size (64) that is used for the
initial buffer and subsequent increases to the buffer. The buffer will
be increased by chunk size until it is large enough to hold the entire
stack of callers.